Move shared_steps and example_data to tests/shared
Also use p: Page type annotation where `tsc` was complaining about it.
