+++ tests/records_other_delivery_site.spec.ts
@@ -1,7 +1,6 @@
-import { test, expect } from "@playwright/test";
-import { examplePatient } from "./example_data";
+import { test, expect, Page } from "@playwright/test";
 
-let p = null;
+let p: Page;
 
 test("Records other delivery site", async ({ page }) => {
   p = page;
